Simplifies the code
