+++ aequilibrae/project/network/link_type.py
@@ -1,31 +1,11 @@
 import string
+from .safe_class import SafeClass
 from aequilibrae.project.database_connection import database_connection
 
 
-class LinkType:
+class LinkType(SafeClass):
     """A link_type object represents a single record in the *link_types* table"""
     __alowed_characters = string.ascii_letters + '_'
-
-    def __init__(self, data_set: dict) -> None:
-        self.__original__ = {}
-        for k, v in data_set.items():
-            self.__dict__[k] = v
-            self.__original__[k] = v
-
-    def __setattr__(self, instance, value) -> None:
-        if instance == 'link_type':
-            if isinstance(value, str):
-                if not len(value):
-                    raise ValueError('link_type cannot be zero-length')
-                for letter in value:
-                    if letter not in self.__alowed_characters:
-                        raise ValueError('link_type can only contain letters and "_"')
-            else:
-                raise ValueError('link_type must be string')
-        if instance == 'link_type_id':
-            raise ValueError('Changing a link_type_id is not supported. Create a new one and delete this')
-        else:
-            self.__dict__[instance] = value
 
     def delete(self):
         conn = database_connection()
@@ -51,3 +31,18 @@ class LinkType:
                     curr.execute(f"update 'link_types' set '{key}'=? where link_type_id='{self.link_type_id}'", [value])
         conn.commit()
         conn.close()
+
+    def __setattr__(self, instance, value) -> None:
+        if instance == 'link_type':
+            if isinstance(value, str):
+                if not len(value):
+                    raise ValueError('link_type cannot be zero-length')
+                for letter in value:
+                    if letter not in self.__alowed_characters:
+                        raise ValueError('link_type can only contain letters and "_"')
+            else:
+                raise ValueError('link_type must be string')
+        if instance == 'link_type_id':
+            raise ValueError('Changing a link_type_id is not supported. Create a new one and delete this')
+        else:
+            self.__dict__[instance] = value
