+++ tests/aequilibrae/project/test_link_type.py
@@ -1,3 +1,4 @@
+from sqlite3 import IntegrityError
 from unittest import TestCase
 import string
 import random
@@ -5,7 +6,7 @@ import os
 from shutil import copytree, rmtree
 import tempfile
 import uuid
-from aequilibrae.project.network import LinkType
+from aequilibrae.project.network.link_type import LinkType
 from aequilibrae.project import Project
 from ...data import no_triggers_project
 
@@ -26,42 +27,34 @@ class TestLinkType(TestCase):
         print(self.temp_proj_folder)
         rmtree(self.temp_proj_folder)
 
-    def test_build(self):
-        for val in ['1', 'ab', '', None]:
-            with self.assertRaises(ValueError):
-                m = LinkType(val)
-
-        for letter in range(10):
-            letter = random.choice(string.ascii_letters)
-            m = LinkType(letter)
-            del m
-
     def test_changing_link_type_id(self):
-        lt = LinkType('X')
+        ltypes = self.proj.network.link_types
+
+        lt = random.choice([x for x in ltypes.all_types().values()])
+
         with self.assertRaises(ValueError):
             lt.link_type_id = 'test my description'
 
-    def test_empty(self):
-        a = LinkType('k')
-        a.link_type = 'just a_test'
         with self.assertRaises(ValueError):
-            a.save()
+            lt.link_type_id = 'K'
 
-        a = LinkType('l')
-        a.link_type = 'just_a_test_test_with_l'
-        with self.assertRaises(ValueError):
-            a.save()
+    def test_empty(self):
+        ltypes = self.proj.network.link_types
+
+        newt = ltypes.new('Z')
+        # a.link_type = 'just a_test'
+        with self.assertRaises(IntegrityError):
+            newt.save()
 
     def test_save(self):
-        self.curr.execute("select link_type_id from 'link_types'")
+        ltypes = self.proj.network.link_types
 
-        letter = random.choice([x[0] for x in self.curr.fetchall()])
-        m = LinkType(letter)
-        m.link_type = self.random_string
-        m.description = self.random_string[::-1]
-        m.save()
+        newt = ltypes.new('Z')
+        newt.link_type = self.random_string
+        newt.description = self.random_string[::-1]
+        newt.save()
 
-        self.curr.execute(f'select description, link_type from link_types where link_type_id="{letter}"')
+        self.curr.execute('select description, link_type from link_types where link_type_id="Z"')
 
         desc, mname = self.curr.fetchone()
         self.assertEqual(desc, self.random_string[::-1], "Didn't save the mode description correctly")
